Detect rename conflicts for uniform names within the same GLSL source chain.
Also add detection for common source chain parts that substitute for GLSL source files in chains that would otherwise be missing a specific shader type.

The fix addresses a rename problem where uniform variable names are global within one shader program and cannot be reused in different shaders. See section 4.3.5 in:
https://www.khronos.org/registry/OpenGL/specs/gl/GLSLangSpec.4.60.pdf
I'd like to thank cce, fizzer and noby for pointing out the issue as compared to the spec.